+++ main.ts
@@ -1,17 +1,18 @@
 import express from "express";
 import { errorMiddleware } from "./src/shared/middleware/error.middleware";
-import { studentsRoutes } from "./src/module/core/students/routes/studentsRoutes.routes";
+import { studentsRoutes } from "./src/module/core/students/routes/students.routes";
 import cookieParser from "cookie-parser";
 import cors from "cors";
-import { authRoutes } from "./src/shared/auth/routes/authRoutes.routes";
+import { authRoutes } from "./src/shared/auth/routes/auth.routes";
 import { usersRoutes } from "./src/module/core/users/routes/users.routes";
 import dotenv from "dotenv";
+import swaggerDocs from "./swagger";
 
 dotenv.config();
 
 export const app = express();
 const router = express.Router();
-const port = 3000;
+const port = 8080;
 
 app.use(cors({ origin: "http://localhost:3000", credentials: true }));
 app.use(express.json());
@@ -31,4 +32,5 @@ app.use(errorMiddleware);
 
 app.listen(port, () => {
   console.log(`Server is running at http://localhost:${port}`);
+  swaggerDocs(app, port);
 });
