+++ package.json
@@ -23,6 +23,8 @@
     "express": "^5.1.0",
     "jsonwebtoken": "^9.0.2",
     "supertest": "^7.1.4",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.1",
     "winston": "^3.17.0"
   },
   "devDependencies": {
@@ -31,6 +33,8 @@
     "@types/jest": "^30.0.0",
     "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "^24.5.2",
+    "@types/swagger-jsdoc": "^6.0.4",
+    "@types/swagger-ui-express": "^4.1.8",
     "jest": "^30.1.3",
     "prisma": "^6.16.2",
     "ts-jest": "^29.4.4",
